Add usec measure from timer0

+++ src/timer0.h
@@ -1,2 +1,4 @@
+#include <stdint.h>
+
 void timer0_initialize(void);
-unsigned long timer0_now_ms(void);
+uint32_t timer0_now_ms(void);
